Documentation: accept ORG and REPO
[ oss commit 1267ff92b0831aa24b1286afcf800b996eb537e6 ]

By default, the check-crd-compat-table script will get the remote from
cilium/cilium. This script won't work if there isn't a remote under
these names. As a workaround, and to avoid many refactoring, the script
will detect if ORG and / or REPO environment variables are set and use
those as inputs to get the remote name.

Signed-off-by: André Martins <andre@cilium.io>
Signed-off-by: Gilberto Bertin <jibi@cilium.io>